Added functionality to register using the API endpoint.

+++ app/src/main/java/com/lifeline/RegistrationData.kt
@@ -1,3 +1,5 @@
 package com.lifeline
 
-data class RegistrationData(val accType:Int, val username:String, val email:String, val id:Int, val password1:String, val password2:String, val agreedToTerms:Boolean=false)
+import java.math.BigInteger
+
+data class RegistrationData(val accType:Int, val username:String, val email:String, val id: BigInteger, val password1:String, val password2:String, val agreedToTerms:Boolean=false)
